% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{algorithms,SeidrFile-method}
\alias{algorithms,SeidrFile-method}
\title{Access algorithm names in a SeidrFile}
\usage{
\S4method{algorithms}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\value{
A character vector of algorithm names
}
\description{
This function returns the algorithm names present
in the \code{SeidrFile} header.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
algorithms(sf)
}
