% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{apply_dummy}
\alias{apply_dummy}
\alias{seidr_apply,SeidrFile-method}
\alias{seidr_chunked_apply,SeidrFile-method}
\alias{seidr_iapply,SeidrFile-method}
\alias{seidr_chunked_iapply,SeidrFile-method}
\title{Vectorized apply functions for SeidrFiles}
\usage{
\S4method{seidr_apply}{SeidrFile}(X, scores = TRUE, ranks = FALSE,
  flags = FALSE, supp_str = FALSE, supp_int = FALSE,
  supp_flt = FALSE, indices = FALSE, FUN, ...)

\S4method{seidr_chunked_apply}{SeidrFile}(X, scores = TRUE,
  ranks = FALSE, flags = FALSE, supp_str = FALSE, supp_int = FALSE,
  supp_flt = FALSE, indices = FALSE, chunksize = 10000, FUN, ...)

\S4method{seidr_iapply}{SeidrFile}(X, scores = TRUE, ranks = FALSE,
  indices = FALSE, index, FUN, ...)

\S4method{seidr_chunked_iapply}{SeidrFile}(X, scores = TRUE,
  ranks = FALSE, indices = FALSE, index, chunksize = 10000, FUN, ...)
}
\arguments{
\item{X}{A \code{\link{SeidrFile-class}} object}

\item{scores}{A \code{logical} indicating if scores should be vectorized}

\item{ranks}{A \code{logical} indicating if ranks should be vectorized}

\item{flags}{A \code{logical} indicating if flags should be vectorized}

\item{supp_str}{A \code{logical} indicating if supplementary strings should
be vectorized}

\item{supp_int}{A \code{logical} indicating if supplementary integers should
be vectorized}

\item{supp_flt}{A \code{logical} indicating if supplementary floats should be
vectorized}

\item{indices}{A \code{logical} indicating if edge indices should be
vectorized}

\item{FUN}{A function to execute}

\item{...}{Additional arguments passed to \code{FUN()}}

\item{chunksize}{The size of a single chunk}

\item{index}{Which score and/or rank to apply \code{FUN()}}
}
\value{
For \code{seidr_apply} and \code{seidr_iapply} the return type is
  whatever \code{FUN()} returns.

  For \code{seidr_chunked_apply} and \code{seidr_chunked_iapply} a list of
  chunks, each containing whatever \code{FUN()} returned for this chunk.
}
\description{
These functions apply \code{FUN()} to entire SeidrFiles in a (mostly)
vectorized manner. \code{FUN()} is passed a vectorized SeidrFile. The
\code{seidr_iapply} and \code{seidr_chunked_iapply} versions apply
\code{FUN()} to a single score and/or rank.
}
\examples{
# Open a connection to a SeidrFile
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)

# Get the best rank for each edge
best_rank <- function(x) {
  return(sapply(x$rank, min, na.rm = TRUE))
}
best <- seidr_apply(sf, scores = FALSE, ranks = TRUE, FUN = best_rank)

# Reset the connection so we can read again
sf <- reset(sf)

# Create a data.frame from a SeidrFile

to_df <- function(x, nodes) {
  data.frame("From"=nodes[x$i], "To"=nodes[x$j],
             "Weight"=x$score, "Rank"=x$rank)
}

df <- seidr_iapply(sf, scores = TRUE, ranks = TRUE, indices = TRUE,
                   index = 10, FUN = to_df, nodes = nodes(sf))

}
