% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{centr_dummy}
\alias{centr_dummy}
\alias{betweenness,SeidrFile-method}
\alias{closeness,SeidrFile-method}
\alias{eigenvector,SeidrFile-method}
\alias{katz,SeidrFile-method}
\alias{strength,SeidrFile-method}
\alias{pagerank,SeidrFile-method}
\title{Access centrality data in a SeidrFile}
\usage{
\S4method{betweenness}{SeidrFile}(object)

\S4method{closeness}{SeidrFile}(object)

\S4method{eigenvector}{SeidrFile}(object)

\S4method{katz}{SeidrFile}(object)

\S4method{strength}{SeidrFile}(object)

\S4method{pagerank}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\value{
A numeric vector of node centrality values
}
\description{
These functions access different node centrality metrics (if computed)
in a SeidrFile. If they weren't computed \code{numeric(0)} is returned.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
betweenness(sf)

}
