% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\name{centrality_accessors}
\alias{centrality_accessors}
\alias{centrality,SeidrFile-method}
\title{Access centrality data in a SeidrFile}
\usage{
\S4method{centrality}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\value{
A numeric vector of node centrality values
}
\description{
These functions access different node centrality metrics (if computed)
in a SeidrFile. If they weren't computed \code{numeric(0)} is returned.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
centrality(sf)

}
