% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{close,SeidrFile-method}
\alias{close,SeidrFile-method}
\title{Close the connection to a SeidrFile}
\usage{
\S4method{close}{SeidrFile}(con, ...)
}
\arguments{
\item{con}{A \code{\link{SeidrFile-class}} object}

\item{...}{Ignored}
}
\description{
Close the connection to a SeidrFile
}
