% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{get_chunk}
\alias{get_chunk}
\alias{get_chunk,SeidrFile-method}
\alias{get_ichunk,SeidrFile-method}
\title{Retrieve one chunk of edge data from a SeidrFile}
\usage{
\S4method{get_chunk}{SeidrFile}(object, chunksize = 10000)

\S4method{get_ichunk}{SeidrFile}(object, score_index, chunksize = 10000)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}

\item{chunksize}{The size of the chunk to retrieve}

\item{score_index}{The index of the score/rank to retrieve}
}
\value{
A list of length \code{chunksize} containing edge data
}
\description{
These functions are mostly meant for development and quick inspection of SeidrFile
contents.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)

get_chunk(sf, chunksize = 10)
}
