% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{nodes,SeidrFile-method}
\alias{nodes,SeidrFile-method}
\title{Access node vector in a SeidrFile}
\usage{
\S4method{nodes}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\value{
A character vector of node names
}
\description{
This function returns the nodes present in the \code{SeidrFile} header.
Edge indices correspond to the order in this vector.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
head(nodes(sf))
}
