% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{reset,SeidrFile-method}
\alias{reset,SeidrFile-method}
\title{Regenerate the connection to a SeidrFile}
\usage{
\S4method{reset}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\description{
This function will reset the SeidrFile for reading.
}
