% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{class}
\name{SeidrFile-class}
\alias{SeidrFile-class}
\title{SeidrFile}
\description{
An R class to access SeidrFiles. It is highly inadvisable for a user to
create an instance of this class with \code{new(...)} unless they
really know what they are doing. Use at your own risk.
}
\section{Slots}{

\describe{
\item{\code{sf_ptr}}{External pointer to SeidrFile C++ interface}

\item{\code{sf_h_ptr}}{External pointer to SeidrFileHeader C++ interface}

\item{\code{sf_attr}}{List of header attributes}

\item{\code{path}}{File path}

\item{\code{i}}{Internal index for source node of current edge}

\item{\code{j}}{Internal index for target node of current edge}

\item{\code{k}}{Internal index for the number of edges read}
}}

