% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\name{SeidrFileFromPath}
\alias{SeidrFileFromPath}
\title{SeidrFileFromPath}
\usage{
SeidrFileFromPath(path = "")
}
\arguments{
\item{path}{The file path of the SeidrFile}
}
\value{
A \code{\link{SeidrFile-class}}
}
\description{
This function creates a new \code{\link{SeidrFile-class}} from a file path. Minimal
checks to ensure the file exists and is indeed a gzip compressed
file.
}
\seealso{
\code{\link{SeidrFile-class}}
}
