% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{show,SeidrFile-method}
\alias{show,SeidrFile-method}
\title{Print SeidrFile metadata}
\usage{
\S4method{show}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\description{
Print SeidrFile metadata
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
show(sf)
}
