% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeidrFile.R
\docType{methods}
\name{supplementary,SeidrFile-method}
\alias{supplementary,SeidrFile-method}
\title{Access supplementary data tags in a SeidrFile}
\usage{
\S4method{supplementary}{SeidrFile}(object)
}
\arguments{
\item{object}{A \code{\link{SeidrFile-class}} object}
}
\value{
A character vector of supplementary data tags.
}
\description{
This function returns the supplementary data tags present
in the \code{SeidrFile} header.
Note that these are only the descriptive tags, for the values one
still needs to unserialize the edge data.
}
\examples{
path <- system.file("extdata", "aggregated.sf", package = "SeidRFile")
sf <- SeidrFileFromPath(path)
supplementary(sf)
}
